requirements.txt
requests>=2.25.1
beautifulsoup4>=4.9.3
pandas>=1.2.0
psycopg2-binary>=2.8.6
python-dotenv>=0.15.0
aiohttp>=3.8.0
beautifulsoup4>=4.9.3
pandas>=1.2.0
asyncpg>=0.27.0
python-dotenv>=0.15.0
aiofiles>=23.1.0